refactor: packages and naming
