+++ composeApp/src/commonMain/kotlin/com/aman/currencyconverter/App.kt
@@ -2,8 +2,8 @@ package com.aman.currencyconverter
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import com.aman.currencyconverter.currencyConverterFeature.presentation.ui.CurrencyConverterScreenRoot
-import com.aman.currencyconverter.currencyConverterFeature.presentation.viewmodel.CurrencyViewModel
+import com.aman.currencyconverter.feature.converter.presentation.ui.CurrencyConverterScreenRoot
+import com.aman.currencyconverter.feature.converter.presentation.viewmodel.CurrencyViewModel
 import org.jetbrains.compose.ui.tooling.preview.Preview
 import org.koin.compose.viewmodel.koinViewModel
