--- composeApp/src/commonMain/kotlin/com/aman/currencyconverter/core/domain/Error.kt
@@ -1,3 +1,0 @@
-package com.aman.currencyconverter.core.domain
-
-interface Error
