+++ composeApp/src/commonMain/kotlin/com/aman/currencyconverter/core/result/AppError.kt
@@ -1,0 +1,3 @@
+package com.aman.currencyconverter.core.result
+
+interface AppError
