+++ composeApp/src/commonMain/kotlin/com/aman/currencyconverter/core/utils/ConverterUtils.kt
@@ -1,46 +1,52 @@
 package com.aman.currencyconverter.core.utils
 
+import kotlin.math.abs
+import kotlin.math.log10
 import kotlin.math.pow
+import kotlin.math.round
 
 
-fun recalculateConvertedAmount(
-    sourceCurrency: String,
-    targetCurrency: String,
-    sourceAmount: String,
-    rates: Map<String, Double>?,
-    onResult: (String) -> Unit
-) {
-    if (rates == null) return
-    val sourceRate = rates[sourceCurrency] ?: return
-    val targetRate = rates[targetCurrency] ?: return
+data class ConversionInput(
+    val sourceCurrency: String,
+    val targetCurrency: String,
+    val amount: String,
+    val rates: Map<String, Double>?
+)
 
-    if (sourceAmount.isBlank()) {
-        onResult("0.0")
-        return
-    }
+/**
+ * Recalculates and returns the converted currency amount as a formatted string.
+ */
+fun convertAmount(input: ConversionInput): String {
+    val rates = input.rates ?: return "0.0"
 
-    val amount = sourceAmount.toDoubleOrNull() ?: return
-    val usd = amount / sourceRate
-    val converted = usd * targetRate
+    val fromRate = rates[input.sourceCurrency] ?: return "0.0"
+    val toRate = rates[input.targetCurrency] ?: return "0.0"
 
-    val formatted = formatToScientificIfLarge(converted, originalInput = sourceAmount)
-    onResult(formatted)
+    val amount = input.amount.toDoubleOrNull() ?: return "0.0"
+
+    // Convert from source currency → USD → target currency
+    val usd = amount / fromRate
+    val converted = usd * toRate
+
+    return formatNicely(converted, input.amount)
 }
 
-fun formatToScientificIfLarge(value: Double, originalInput: String, maxDecimals: Int = 4): String {
+
+/**
+ * Formats the value:
+ * - Uses scientific format if input is 6+ digits
+ * - Otherwise rounds to 2 decimals
+ */
+private fun formatNicely(value: Double, originalInput: String): String {
     if (value == 0.0) return "0.0"
 
     return if (originalInput.length >= 6) {
-        val exponent = kotlin.math.floor(kotlin.math.log10(kotlin.math.abs(value))).toInt()
-        val mantissa = value / 10.0.pow(exponent)
-
-        // Round mantissa to required decimal places
-        val multiplier = 10.0.pow(maxDecimals)
-        val roundedMantissa = kotlin.math.round(mantissa * multiplier) / multiplier
-
-        "${roundedMantissa}e${if (exponent >= 0) "+" else ""}$exponent"
+        val exponent = log10(abs(value)).toInt()
+        val base = value / 10.0.pow(exponent)
+        val roundedBase = (round(base * 10000) / 10000)
+        "${roundedBase}e${if (exponent >= 0) "+" else ""}$exponent"
     } else {
-        val rounded = kotlin.math.round(value * 100) / 100
+        val rounded = (round(value * 100) / 100)
         rounded.toString()
     }
 }
