--- composeApp/src/commonMain/kotlin/com/aman/currencyconverter/currencyConverterFeature/domain/repository/CurrencyExchangeRepository.kt
@@ -1,9 +1,0 @@
-package com.aman.currencyconverter.currencyConverterFeature.domain.repository
-
-import com.aman.currencyconverter.core.domain.DataError
-import com.aman.currencyconverter.core.domain.Result
-import com.aman.currencyconverter.currencyConverterFeature.domain.model.ExchangeRate
-
-interface CurrencyExchangeRepository {
-    suspend fun fetchExchangeRates(): Result<ExchangeRate, DataError.Remote>
-}
