--- composeApp/src/commonMain/kotlin/com/aman/currencyconverter/currencyConverterFeature/presentation/ui/components/CurrencyDropdown.kt
@@ -1,67 +1,0 @@
-package com.aman.currencyconverter.currencyConverterFeature.presentation.ui.components
-
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowDropDown
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun CurrencyDropdown(
-    selectedCurrency: String,
-    currencies: List<String>,
-    onCurrencySelected: (String) -> Unit
-) {
-    var expanded by remember { mutableStateOf(false) }
-
-    Box {
-        Row(
-            modifier = Modifier
-                .clickable { expanded = true }
-                .padding(vertical = 8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Text(
-                text = selectedCurrency,
-                color = Color.White,
-                style = MaterialTheme.typography.bodyLarge
-            )
-            Icon(
-                imageVector = Icons.Default.ArrowDropDown,
-                contentDescription = "Dropdown Arrow",
-                tint = Color.White
-            )
-        }
-
-        DropdownMenu(
-            expanded = expanded,
-            onDismissRequest = { expanded = false }
-        ) {
-            currencies.forEach { currency ->
-                DropdownMenuItem(
-                    text = { Text(currency) },
-                    onClick = {
-                        onCurrencySelected(currency)
-                        expanded = false
-                    }
-                )
-            }
-        }
-    }
-}
