+++ composeApp/src/commonMain/kotlin/com/aman/currencyconverter/feature/converter/domain/repository/CurrencyExchangeRepository.kt
@@ -1,0 +1,9 @@
+package com.aman.currencyconverter.feature.converter.domain.repository
+
+import com.aman.currencyconverter.core.result.DataFetchAppError
+import com.aman.currencyconverter.core.result.AppResult
+import com.aman.currencyconverter.feature.converter.domain.model.ExchangeRate
+
+interface CurrencyExchangeRepository {
+    suspend fun fetchExchangeRates(): AppResult<ExchangeRate, DataFetchAppError.Remote>
+}
